# 03_api_gateway/api/app/schemas/users.py
"""
File: /app/03_api_gateway/api/app/schemas/users.py
x-lucid-file-path: /app/03_api_gateway/api/app/schemas/users.py
x-lucid-file-directory: /app/03_api_gateway/api/app/schemas
x-lucid-file-type: python
"""

from pydantic import BaseModel, EmailStr, Field
from typing import List
from datetime import datetime


class UserPublic(BaseModel):
    id: str = Field()
    email: EmailStr
    role: str = Field()
    created_at: datetime
    updated_at: datetime


class RegisterRequest(BaseModel):
    email: EmailStr
    password: str = Field(min_length=8, max_length=128)


class PaginatedUsers(BaseModel):
    items: List[UserPublic]
    page: int
    page_size: int
    total: int